add somethings

+++ Data/Entities/Question.cs
@@ -15,7 +15,7 @@ namespace Data.Entities
         public string? QC { get; set; }
         public string? QD { get; set; }
         public string? QCorrectAns { get; set; }
-        public int? Score { get; set; }
+        public float? Score { get; set; }
         public int CategoryId { get; set; }
         public Category Category { get; set; }
     }
